Simplify several functions in pre-processing
I am uncertain of the quality and speed. There are probably still
fundamental speed issues compared to using pointer.

Ether tree is not built correctly. It is happening faster now which
could be the effect of some optimized pre processing functions.

The query effectiveness is copmletely wasted. Perhaps because pointers
are not used there anymore.